Form, tabs and style, as close as possible to dblog code

+++ modules/mongodb_watchdog/src/Controller/AdminController.php
@@ -8,20 +8,23 @@
 namespace Drupal\mongodb_watchdog\Controller;
 
 use Drupal\Component\Utility\Unicode;
-use Drupal\Core\DependencyInjection\ContainerInjectionInterface;
+use Drupal\Core\Controller\ControllerBase;
 use Drupal\Core\Url;
 use Drupal\mongodb_watchdog\Logger;
 use MongoDB\Database;
 use Psr\Log\LogLevel;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+use Drupal\Core\Extension\ModuleHandlerInterface;
+use Drupal\Core\Form\FormBuilderInterface;
+
 
 /**
  * Class AdminController.
  *
  * @package Drupal\mongodb_watchdog
  */
-class AdminController implements ContainerInjectionInterface {
+class AdminController extends ControllerBase {
 
   /**
    * The MongoDB database for the logger alias.
@@ -44,20 +47,38 @@ class AdminController implements ContainerInjectionInterface {
    */
   protected $watchdog;
 
+   * The module handler service.
+   *
+   * @var \Drupal\Core\Extension\ModuleHandlerInterface
+   */
+  protected $moduleHandler;
+
+  /**
+   * The form builder service.
+   *
+   * @var \Drupal\Core\Form\FormBuilderInterface
+   */
+  protected $formBuilder;
+
   /**
    * Constructor.
    *
-   * @param \MongoDB\Database $database
+   * @param \MongoDB $database
    *   The watchdog database.
    * @param \Psr\Log\LoggerInterface $logger
    *   The logger service, to log intervening events.
    * @param \Drupal\mongodb_watchdog\Logger $watchdog
    *   The MongoDB logger, to load stored events.
+   * @param \Drupal\Core\Extension\ModuleHandlerInterface $module_handler
+   *   A module handler.
+   * @param \Drupal\Core\Form\FormBuilderInterface $form_builder
+   *   The form builder service.
    */
-  public function __construct(Database $database, LoggerInterface $logger, Logger $watchdog) {
+  public function __construct(Database $database, LoggerInterface $logger, Logger $watchdog, ModuleHandlerInterface $module_handler, FormBuilderInterface $form_builder) {
     $this->database = $database;
     $this->logger = $logger;
-    $this->watchdog = $watchdog;
+    $this->moduleHandler = $module_handler;
+    $this->formBuilder = $form_builder;
   }
 
   /**
@@ -81,6 +102,10 @@ class AdminController implements ContainerInjectionInterface {
     $collection = $this->database->selectCollection(Logger::TEMPLATE_COLLECTION);
     $cursor = $collection->find();
 
+    $this->moduleHandler->loadInclude('mongodb_watchdog', 'admin.inc');
+
+    $build['dblog_filter_form'] = $this->formBuilder->getForm('Drupal\mongodb_watchdog\Form\MongodbWatchdogFilterForm');
+
     $header = array(
       // Icon column.
       '',
@@ -93,41 +118,47 @@ class AdminController implements ContainerInjectionInterface {
 
     $rows = array();
     foreach ($cursor as $id => $value) {
-      // dsm($value, $id);
-      /*
-      if ($value['type'] == 'php' && $value['message'] == '%type: %message in %function (line %line of %file).') {
-        $collection = $this->logger->eventCollection($value['_id']);
-        $result = $collection->find()
-                             ->sort(array('$natural' => -1))
-                             ->limit(1)
-                             ->getNext();
-        if ($value) {
-          $value['file'] = basename($result['variables']['%file']);
-          $value['line'] = $result['variables']['%line'];
-          $value['message'] = '%type in %function';
-          $value['variables'] = $result['variables'];
-        }
+      if ($id < 5) {
+        kint($value, $id);
+        kint($value->_id);
+//        if ($value['type'] == 'php' && $value['message'] == '%type: %message in %function (line %line of %file).') {
+//          $collection = $this->logger->eventCollection($value['_id']);
+//          $result = $collection->find()
+//            ->sort(['$natural' => -1])
+//            ->limit(1)
+//            ->getNext();
+//          if ($value) {
+//            $value['file'] = basename($result['variables']['%file']);
+//            $value['line'] = $result['variables']['%line'];
+//            $value['message'] = '%type in %function';
+//            $value['variables'] = $result['variables'];
+//          }
+//        }
+        $message = Unicode::truncate(strip_tags(SafeMarkup::format($value['message'], [])), 56, TRUE, TRUE);
+        $value['count'] = 0; // $this->logger->eventCollection($value['_id'])->count();
+        $rows[$id] = [
+          $icons[$value->severity],
+          isset($value['count']) && $value['count'] > 1 ? intval($value['count']) : 0,
+          t($value['type']),
+          empty($value['timestamp']) ? '' : format_date($value['timestamp'], 'short'),
+          empty($value['file']) ? '' : Unicode::truncate(basename($value['file']), 30) . (empty($value['line']) ? '' : ('+' . $value['line'])),
+          \Drupal::l($message, Url::fromRoute('mongodb_watchdog.reports.detail', ['id' => $id])),
+        ];
       }
-      */
-      $message = ''; //Unicode::truncate(strip_tags(SafeMarkup::format($value)), 56, TRUE, TRUE);
-      //$value['count'] = $this->logger->eventCollection($value['_id'])->count();
-      $rows[$id] = array(
-        //$icons[$value['severity']],
-        isset($value['count']) && $value['count'] > 1 ? intval($value['count']) : 0,
-        t($value['type']),
-        empty($value['timestamp']) ? '' : format_date($value['timestamp'], 'short'),
-        empty($value['file']) ? '' : Unicode::truncate(basename($value['file']), 30) . (empty($value['line']) ? '' : ('+' . $value['line'])),
-        \Drupal::l($message, Url::fromRoute('mongodb_watchdog.reports.detail', ['event_template' => $id])),
-      );
-    }
 
+    }
+kint($rows);
     $build['mongodb_watchdog_table'] = array(
       '#theme' => 'table',
       '#header' => $header,
       '#rows' => $rows,
       '#attributes' => ['id' => 'admin-mongodb_watchdog'],
+      '#attached' => array(
+        'library' => array('mongodb_watchdog/drupal.mongodb_watchdog'),
+      ),
     );
 
+    $build['mongodb_watchdog_pager'] = array('#type' => 'pager');
 
     return $build;
   }
@@ -145,12 +176,16 @@ class AdminController implements ContainerInjectionInterface {
     /** @var \MongoDB $database */
     $database = $container->get('mongodb.watchdog_storage');
 
+    $form_builder = $container->get('form_builder');
+
+    $module_handler = $container->get('module_handler');
+
     /** @var \Psr\Log\LoggerInterface $logger */
     $logger = $container->get('logger.channel.mongodb_watchdog');
 
     /** @var \Drupal\mongodb_watchdog\Logger $logger */
     $watchdog = $container->get('mongodb.logger');
 
-    return new static($database, $logger, $watchdog);
+    return new static($database, $logger, $watchdog, $module_handler, $form_builder);
   }
 }
